feat(register): add extensions file .
