+++ lib/core/utilities/extensions/gender_ex.dart
@@ -1,0 +1,12 @@
+enum Gender { male, female }
+
+extension GenderExtension on Gender {
+  String getValue() {
+    switch (this) {
+      case Gender.male:
+        return 'male';
+      case Gender.female:
+        return 'female';
+    }
+  }
+}
